--- a/packages/angular/cli/bin/bootstrap.js
+++ b/packages/angular/cli/bin/bootstrap.js
@@ -18,4 +18,16 @@
  * range.
  */
 
-import('../lib/init.js');
+// Enable on-disk code caching if available (Node.js 22.8+)
+// Skip if running inside Bazel via a RUNFILES environment variable check. The cache does not work
+// well with Bazel's hermeticity requirements.
+if (!process.env['RUNFILES']) {
+  try {
+    const { enableCompileCache } = require('node:module');
+
+    enableCompileCache?.();
+  } catch {}
+}
+
+// Initialize the Angular CLI
+void import('../lib/init.js');